+++ changelog.md
@@ -6,6 +6,7 @@
 - ADDED `install` with installation folder and subfolder resolution
 - FIXED documentation from `@return` to `@returns`
 - REFACTORED the `uri` is parsed to see if the version is embedded in the archive name
+- ADDED logging to `vault.log` until the console messaging is clearer
 
 ## 0.4.0